requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
jupyter>=1.0.0
tqdm>=4.65.0
